Fix wrong words and ordering for staff member form (#1424)
* fix wrong words and ordering for staff member form

* add messages

* remove line

* apply changes from review

* change to local messages
