--- src/components/AppStatus/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./AppStatus";
-export * from "./AppStatus";
