+++ src/staff/components/StaffDetailsPage/StaffDetailsPage.tsx
@@ -1,6 +1,5 @@
 import { Card, CardContent, Typography } from "@material-ui/core";
 import AccountPermissionGroups from "@saleor/components/AccountPermissionGroups";
-import AccountStatus from "@saleor/components/AppStatus";
 import CardSpacer from "@saleor/components/CardSpacer";
 import CardTitle from "@saleor/components/CardTitle";
 import { ConfirmButtonTransitionState } from "@saleor/components/ConfirmButton";
@@ -18,6 +17,7 @@ import { sectionNames } from "@saleor/intl";
 import { Backlink } from "@saleor/macaw-ui";
 import { getUserName } from "@saleor/misc";
 import { SearchPermissionGroups_search_edges_node } from "@saleor/searches/types/SearchPermissionGroups";
+import UserStatus from "@saleor/staff/components/UserStatus";
 import { FetchMoreProps, SearchPageProps } from "@saleor/types";
 import createMultiAutocompleteSelectHandler from "@saleor/utils/handlers/multiAutocompleteSelectChangeHandler";
 import React from "react";
@@ -27,6 +27,7 @@ import { StaffMemberDetails_user } from "../../types/StaffMemberDetails";
 import StaffPassword from "../StaffPassword/StaffPassword";
 import StaffPreferences from "../StaffPreferences";
 import StaffProperties from "../StaffProperties/StaffProperties";
+import { staffDetailsPageMessages as messages } from "./messages";
 
 export interface StaffDetailsFormData {
   email: string;
@@ -143,6 +144,13 @@ const StaffDetailsPage: React.FC<StaffDetailsPageProps> = ({
                 )}
                 {canEditStatus && (
                   <>
+                    <UserStatus
+                      data={formData}
+                      disabled={disabled}
+                      label={intl.formatMessage(messages.userStatusActive)}
+                      onChange={change}
+                    />
+                    <CardSpacer />
                     <Card>
                       <CardTitle
                         title={intl.formatMessage({
@@ -171,16 +179,6 @@ const StaffDetailsPage: React.FC<StaffDetailsPageProps> = ({
                         />
                       </CardContent>
                     </Card>
-                    <CardSpacer />
-                    <AccountStatus
-                      data={formData}
-                      disabled={disabled}
-                      label={intl.formatMessage({
-                        defaultMessage: "User is active",
-                        description: "checkbox label"
-                      })}
-                      onChange={change}
-                    />
                   </>
                 )}
               </div>
