+++ src/staff/components/StaffDetailsPage/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from "react-intl";
+
+export const staffDetailsPageMessages = defineMessages({
+  userStatusActive: {
+    defaultMessage: "User is active",
+    description: "checkbox label"
+  }
+});
