+++ src/staff/components/UserStatus/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from "react-intl";
+
+export const userStatusMessages = defineMessages({
+  userStatus: {
+    defaultMessage: "User Status",
+    description: "section header"
+  },
+  userDisableInstruction: {
+    defaultMessage:
+      "If you want to disable this User please uncheck the box below."
+  }
+});
